msgs: move handshake payload types into msgs
Helps avoid circular imports between the _hello modules and handshake.
