+++ rustls/src/msgs/handshake_test.rs
@@ -17,17 +17,17 @@ use super::enums::{
 use super::handshake::{
     CertificateChain, CertificateEntry, CertificateExtensions, CertificatePayloadTls13,
     CertificateRequestExtensions, CertificateRequestPayload, CertificateRequestPayloadTls13,
-    CertificateStatus, CompressedCertificatePayload, EcParameters, HandshakeMessagePayload,
-    HandshakePayload, HelloRetryRequest, HelloRetryRequestExtensions, KeyShareEntry,
-    NewSessionTicketExtensions, NewSessionTicketPayload, NewSessionTicketPayloadTls13, Random,
-    ServerDhParams, ServerEcdhParams, ServerKeyExchange, ServerKeyExchangeParams,
-    ServerKeyExchangePayload, SessionId, SingleProtocolName, SupportedEcPointFormats,
-    SupportedProtocolVersions,
+    CertificateStatus, CompressedCertificatePayload, EcParameters, HelloRetryRequest,
+    HelloRetryRequestExtensions, KeyShareEntry, NewSessionTicketExtensions,
+    NewSessionTicketPayload, NewSessionTicketPayloadTls13, Random, ServerDhParams,
+    ServerEcdhParams, ServerKeyExchange, ServerKeyExchangeParams, ServerKeyExchangePayload,
+    SessionId, SingleProtocolName, SupportedEcPointFormats, SupportedProtocolVersions,
 };
 use super::server_hello::{
     EchConfigContents, EchConfigPayload, HpkeKeyConfig, ServerEncryptedClientHello,
     ServerExtensions, ServerHelloPayload,
 };
+use super::{HandshakeMessagePayload, HandshakePayload};
 use crate::crypto::cipher::Payload;
 use crate::crypto::hpke::{HpkeAead, HpkeKdf, HpkeKem, HpkeSymmetricCipherSuite};
 use crate::crypto::kx::NamedGroup;
