+++ rustls/src/msgs/mod.rs
@@ -31,11 +31,13 @@
 //!
 //! <https://langsec.org/ForWantOfANail-h2hc2014.pdf>
 
+use alloc::boxed::Box;
 use alloc::vec::Vec;
 
 use crate::crypto::cipher::{EncodedMessage, MessageError, Payload};
 use crate::enums::{ContentType, HandshakeType, ProtocolVersion};
 use crate::error::{AlertDescription, InvalidMessage};
+use crate::verify::DigitallySignedStruct;
 
 #[macro_use]
 mod macros;
@@ -53,6 +55,7 @@ pub(crate) use client_hello::{
 mod codec;
 pub(crate) use codec::{CERTIFICATE_MAX_SIZE_LIMIT, ListLength, TlsListElement, put_u16, put_u64};
 pub use codec::{Codec, Reader};
+use codec::{LengthPrefixedBuffer, U24};
 
 mod deframer;
 pub use deframer::fuzz_deframer;
@@ -75,15 +78,16 @@ pub use fragmenter::MessageFragmenter;
 
 #[macro_use]
 mod handshake;
+use handshake::HELLO_RETRY_REQUEST_RANDOM;
 pub(crate) use handshake::{
     ALL_KEY_EXCHANGE_ALGORITHMS, CertificateChain, CertificatePayloadTls13,
     CertificateRequestExtensions, CertificateRequestPayload, CertificateRequestPayloadTls13,
     CertificateStatus, ClientDhParams, ClientEcdhParams, ClientExtensionsInput,
-    ClientKeyExchangeParams, CompressedCertificatePayload, Encoding, HandshakeMessagePayload,
-    HandshakePayload, HelloRetryRequest, HelloRetryRequestExtensions, KeyShareEntry, KxDecode,
-    NewSessionTicketPayload, NewSessionTicketPayloadTls13, Random, ServerExtensionsInput,
-    ServerKeyExchange, ServerKeyExchangeParams, ServerKeyExchangePayload, SessionId,
-    SingleProtocolName, SupportedEcPointFormats, SupportedProtocolVersions, TransportParameters,
+    ClientKeyExchangeParams, CompressedCertificatePayload, Encoding, HelloRetryRequest,
+    HelloRetryRequestExtensions, KeyShareEntry, KxDecode, NewSessionTicketPayload,
+    NewSessionTicketPayloadTls13, Random, ServerExtensionsInput, ServerKeyExchange,
+    ServerKeyExchangeParams, ServerKeyExchangePayload, SessionId, SingleProtocolName,
+    SupportedEcPointFormats, SupportedProtocolVersions, TransportParameters,
 };
 #[cfg(test)]
 pub(crate) use handshake::{EcParameters, ServerEcdhParams};
@@ -322,6 +326,294 @@ impl From<Message<'_>> for EncodedMessage<Payload<'_>> {
 }
 
 #[derive(Debug)]
+pub struct HandshakeMessagePayload<'a>(pub(crate) HandshakePayload<'a>);
+
+impl<'a> Codec<'a> for HandshakeMessagePayload<'a> {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        self.payload_encode(bytes, Encoding::Standard);
+    }
+
+    fn read(r: &mut Reader<'a>) -> Result<Self, InvalidMessage> {
+        Self::read_version(r, ProtocolVersion::TLSv1_2)
+    }
+}
+
+impl<'a> HandshakeMessagePayload<'a> {
+    pub(crate) fn read_version(
+        r: &mut Reader<'a>,
+        vers: ProtocolVersion,
+    ) -> Result<Self, InvalidMessage> {
+        let typ = HandshakeType::read(r)?;
+        let len = U24::read(r)?.0 as usize;
+        let mut sub = r.sub(len)?;
+
+        let payload = match typ {
+            HandshakeType::HelloRequest if sub.left() == 0 => HandshakePayload::HelloRequest,
+            HandshakeType::ClientHello => {
+                HandshakePayload::ClientHello(ClientHelloPayload::read(&mut sub)?)
+            }
+            HandshakeType::ServerHello => {
+                let version = ProtocolVersion::read(&mut sub)?;
+                let random = Random::read(&mut sub)?;
+
+                if random == HELLO_RETRY_REQUEST_RANDOM {
+                    let mut hrr = HelloRetryRequest::read(&mut sub)?;
+                    hrr.legacy_version = version;
+                    HandshakePayload::HelloRetryRequest(hrr)
+                } else {
+                    let mut shp = ServerHelloPayload::read(&mut sub)?;
+                    shp.legacy_version = version;
+                    shp.random = random;
+                    HandshakePayload::ServerHello(shp)
+                }
+            }
+            HandshakeType::Certificate if vers == ProtocolVersion::TLSv1_3 => {
+                let p = CertificatePayloadTls13::read(&mut sub)?;
+                HandshakePayload::CertificateTls13(p)
+            }
+            HandshakeType::Certificate => {
+                HandshakePayload::Certificate(CertificateChain::read(&mut sub)?)
+            }
+            HandshakeType::ServerKeyExchange => {
+                let p = ServerKeyExchangePayload::read(&mut sub)?;
+                HandshakePayload::ServerKeyExchange(p)
+            }
+            HandshakeType::ServerHelloDone => {
+                sub.expect_empty("ServerHelloDone")?;
+                HandshakePayload::ServerHelloDone
+            }
+            HandshakeType::ClientKeyExchange => {
+                HandshakePayload::ClientKeyExchange(Payload::read(&mut sub))
+            }
+            HandshakeType::CertificateRequest if vers == ProtocolVersion::TLSv1_3 => {
+                let p = CertificateRequestPayloadTls13::read(&mut sub)?;
+                HandshakePayload::CertificateRequestTls13(p)
+            }
+            HandshakeType::CertificateRequest => {
+                let p = CertificateRequestPayload::read(&mut sub)?;
+                HandshakePayload::CertificateRequest(p)
+            }
+            HandshakeType::CompressedCertificate => HandshakePayload::CompressedCertificate(
+                CompressedCertificatePayload::read(&mut sub)?,
+            ),
+            HandshakeType::CertificateVerify => {
+                HandshakePayload::CertificateVerify(DigitallySignedStruct::read(&mut sub)?)
+            }
+            HandshakeType::NewSessionTicket if vers == ProtocolVersion::TLSv1_3 => {
+                let p = NewSessionTicketPayloadTls13::read(&mut sub)?;
+                HandshakePayload::NewSessionTicketTls13(p)
+            }
+            HandshakeType::NewSessionTicket => {
+                let p = NewSessionTicketPayload::read(&mut sub)?;
+                HandshakePayload::NewSessionTicket(p)
+            }
+            HandshakeType::EncryptedExtensions => {
+                HandshakePayload::EncryptedExtensions(Box::new(ServerExtensions::read(&mut sub)?))
+            }
+            HandshakeType::KeyUpdate => {
+                HandshakePayload::KeyUpdate(KeyUpdateRequest::read(&mut sub)?)
+            }
+            HandshakeType::EndOfEarlyData => {
+                sub.expect_empty("EndOfEarlyData")?;
+                HandshakePayload::EndOfEarlyData
+            }
+            HandshakeType::Finished => HandshakePayload::Finished(Payload::read(&mut sub)),
+            HandshakeType::CertificateStatus => {
+                HandshakePayload::CertificateStatus(CertificateStatus::read(&mut sub)?)
+            }
+            HandshakeType::MessageHash => {
+                // does not appear on the wire
+                return Err(InvalidMessage::UnexpectedMessage("MessageHash"));
+            }
+            HandshakeType::HelloRetryRequest => {
+                // not legal on wire
+                return Err(InvalidMessage::UnexpectedMessage("HelloRetryRequest"));
+            }
+            _ => HandshakePayload::Unknown((typ, Payload::read(&mut sub))),
+        };
+
+        sub.expect_empty("HandshakeMessagePayload")
+            .map(|_| Self(payload))
+    }
+
+    pub(crate) fn encoding_for_binder_signing(&self) -> Vec<u8> {
+        let mut ret = self.get_encoding();
+        let ret_len = ret.len() - self.total_binder_length();
+        ret.truncate(ret_len);
+        ret
+    }
+
+    pub(crate) fn total_binder_length(&self) -> usize {
+        match &self.0 {
+            HandshakePayload::ClientHello(ch) => match &ch.preshared_key_offer {
+                Some(offer) => {
+                    let mut binders_encoding = Vec::new();
+                    offer
+                        .binders
+                        .encode(&mut binders_encoding);
+                    binders_encoding.len()
+                }
+                _ => 0,
+            },
+            _ => 0,
+        }
+    }
+
+    pub(crate) fn payload_encode(&self, bytes: &mut Vec<u8>, encoding: Encoding) {
+        // output type, length, and encoded payload
+        self.0
+            .wire_handshake_type()
+            .encode(bytes);
+
+        let nested = LengthPrefixedBuffer::new(
+            ListLength::U24 {
+                max: usize::MAX,
+                error: InvalidMessage::MessageTooLarge,
+            },
+            bytes,
+        );
+
+        match &self.0 {
+            // for Server Hello and HelloRetryRequest payloads we need to encode the payload
+            // differently based on the purpose of the encoding.
+            HandshakePayload::ServerHello(payload) => payload.payload_encode(nested.buf, encoding),
+            HandshakePayload::HelloRetryRequest(payload) => {
+                payload.payload_encode(nested.buf, encoding)
+            }
+
+            // All other payload types are encoded the same regardless of purpose.
+            _ => self.0.encode(nested.buf),
+        }
+    }
+
+    pub(crate) fn build_handshake_hash(hash: &[u8]) -> Self {
+        Self(HandshakePayload::MessageHash(Payload::new(hash.to_vec())))
+    }
+
+    #[cfg(feature = "std")]
+    pub(crate) fn into_owned(self) -> HandshakeMessagePayload<'static> {
+        HandshakeMessagePayload(self.0.into_owned())
+    }
+}
+
+#[derive(Debug)]
+pub(crate) enum HandshakePayload<'a> {
+    HelloRequest,
+    ClientHello(ClientHelloPayload),
+    ServerHello(ServerHelloPayload),
+    HelloRetryRequest(HelloRetryRequest),
+    Certificate(CertificateChain<'a>),
+    CertificateTls13(CertificatePayloadTls13<'a>),
+    CompressedCertificate(CompressedCertificatePayload<'a>),
+    ServerKeyExchange(ServerKeyExchangePayload),
+    CertificateRequest(CertificateRequestPayload),
+    CertificateRequestTls13(CertificateRequestPayloadTls13),
+    CertificateVerify(DigitallySignedStruct),
+    ServerHelloDone,
+    EndOfEarlyData,
+    ClientKeyExchange(Payload<'a>),
+    NewSessionTicket(NewSessionTicketPayload),
+    NewSessionTicketTls13(NewSessionTicketPayloadTls13),
+    EncryptedExtensions(Box<ServerExtensions<'a>>),
+    KeyUpdate(KeyUpdateRequest),
+    Finished(Payload<'a>),
+    CertificateStatus(CertificateStatus<'a>),
+    MessageHash(Payload<'a>),
+    Unknown((HandshakeType, Payload<'a>)),
+}
+
+impl HandshakePayload<'_> {
+    fn encode(&self, bytes: &mut Vec<u8>) {
+        use self::HandshakePayload::*;
+        match self {
+            HelloRequest | ServerHelloDone | EndOfEarlyData => {}
+            ClientHello(x) => x.encode(bytes),
+            ServerHello(x) => x.encode(bytes),
+            HelloRetryRequest(x) => x.encode(bytes),
+            Certificate(x) => x.encode(bytes),
+            CertificateTls13(x) => x.encode(bytes),
+            CompressedCertificate(x) => x.encode(bytes),
+            ServerKeyExchange(x) => x.encode(bytes),
+            ClientKeyExchange(x) => x.encode(bytes),
+            CertificateRequest(x) => x.encode(bytes),
+            CertificateRequestTls13(x) => x.encode(bytes),
+            CertificateVerify(x) => x.encode(bytes),
+            NewSessionTicket(x) => x.encode(bytes),
+            NewSessionTicketTls13(x) => x.encode(bytes),
+            EncryptedExtensions(x) => x.encode(bytes),
+            KeyUpdate(x) => x.encode(bytes),
+            Finished(x) => x.encode(bytes),
+            CertificateStatus(x) => x.encode(bytes),
+            MessageHash(x) => x.encode(bytes),
+            Unknown((_, x)) => x.encode(bytes),
+        }
+    }
+
+    pub(crate) fn handshake_type(&self) -> HandshakeType {
+        use self::HandshakePayload::*;
+        match self {
+            HelloRequest => HandshakeType::HelloRequest,
+            ClientHello(_) => HandshakeType::ClientHello,
+            ServerHello(_) => HandshakeType::ServerHello,
+            HelloRetryRequest(_) => HandshakeType::HelloRetryRequest,
+            Certificate(_) | CertificateTls13(_) => HandshakeType::Certificate,
+            CompressedCertificate(_) => HandshakeType::CompressedCertificate,
+            ServerKeyExchange(_) => HandshakeType::ServerKeyExchange,
+            CertificateRequest(_) | CertificateRequestTls13(_) => HandshakeType::CertificateRequest,
+            CertificateVerify(_) => HandshakeType::CertificateVerify,
+            ServerHelloDone => HandshakeType::ServerHelloDone,
+            EndOfEarlyData => HandshakeType::EndOfEarlyData,
+            ClientKeyExchange(_) => HandshakeType::ClientKeyExchange,
+            NewSessionTicket(_) | NewSessionTicketTls13(_) => HandshakeType::NewSessionTicket,
+            EncryptedExtensions(_) => HandshakeType::EncryptedExtensions,
+            KeyUpdate(_) => HandshakeType::KeyUpdate,
+            Finished(_) => HandshakeType::Finished,
+            CertificateStatus(_) => HandshakeType::CertificateStatus,
+            MessageHash(_) => HandshakeType::MessageHash,
+            Unknown((t, _)) => *t,
+        }
+    }
+
+    fn wire_handshake_type(&self) -> HandshakeType {
+        match self.handshake_type() {
+            // A `HelloRetryRequest` appears on the wire as a `ServerHello` with a magic `random` value.
+            HandshakeType::HelloRetryRequest => HandshakeType::ServerHello,
+            other => other,
+        }
+    }
+
+    #[cfg(feature = "std")]
+    fn into_owned(self) -> HandshakePayload<'static> {
+        use HandshakePayload::*;
+
+        match self {
+            HelloRequest => HelloRequest,
+            ClientHello(x) => ClientHello(x),
+            ServerHello(x) => ServerHello(x),
+            HelloRetryRequest(x) => HelloRetryRequest(x),
+            Certificate(x) => Certificate(x.into_owned()),
+            CertificateTls13(x) => CertificateTls13(x.into_owned()),
+            CompressedCertificate(x) => CompressedCertificate(x.into_owned()),
+            ServerKeyExchange(x) => ServerKeyExchange(x),
+            CertificateRequest(x) => CertificateRequest(x),
+            CertificateRequestTls13(x) => CertificateRequestTls13(x),
+            CertificateVerify(x) => CertificateVerify(x),
+            ServerHelloDone => ServerHelloDone,
+            EndOfEarlyData => EndOfEarlyData,
+            ClientKeyExchange(x) => ClientKeyExchange(x.into_owned()),
+            NewSessionTicket(x) => NewSessionTicket(x),
+            NewSessionTicketTls13(x) => NewSessionTicketTls13(x),
+            EncryptedExtensions(x) => EncryptedExtensions(Box::new(x.into_owned())),
+            KeyUpdate(x) => KeyUpdate(x),
+            Finished(x) => Finished(x.into_owned()),
+            CertificateStatus(x) => CertificateStatus(x.into_owned()),
+            MessageHash(x) => MessageHash(x.into_owned()),
+            Unknown((t, x)) => Unknown((t, x.into_owned())),
+        }
+    }
+}
+
+#[derive(Debug)]
 pub struct AlertMessagePayload {
     pub level: AlertLevel,
     pub description: AlertDescription,
@@ -378,7 +670,6 @@ mod tests {
     use std::path::{Path, PathBuf};
     use std::{format, fs, println};
 
-    use super::codec::U24;
     use super::*;
     use crate::crypto::cipher::OutboundOpaque;
